refactor: eliminate duplicated StorageInsertionResult enum
Move TryFrom<u8> impl to zenoh-backend-traits (source of truth) and
replace the duplicate enum in zenoh-ext with a pub use re-export.
Remove the discriminant parity test that is no longer needed.

Closes #123
