+++ zenoh-ext/Cargo.toml
@@ -48,12 +48,12 @@ tokio = { workspace = true, features = [
 tracing = { workspace = true }
 uhlc = { workspace = true }
 zenoh = { workspace = true, default-features = false }
+zenoh_backend_traits = { workspace = true }
 zenoh-macros = { workspace = true }
 zenoh-util = { workspace = true }
 
 [dev-dependencies]
 rand = { workspace = true }
-zenoh_backend_traits = { workspace = true }
 zenoh-config = { workspace = true }
 zenoh-plugin-storage-manager = { path = "../plugins/zenoh-plugin-storage-manager" }
 zenoh-plugin-trait = { workspace = true }
